Correct use of $stmt->fetch();
Was not using $stmt->fetch(); correctly.  Now it should be working
correctly.  Was expecting it to return an array of results, it returns a
code related to the successs or failure of the querry. 
$stmt->bind_result($result1, ...) is required to get the results, fetch
returns one result to the bound variables.  Multiple results must be
feched through multiple calls to fetch().