Create User setup as funcion
controller/createuser.php - Changed code from old stand alone
controller, to the ability to work as part of the universal controller,
with the old code being converted to a function.

Fixed the require to a require_once for the debug code.  May have to do
that for other code as well.

To Do - Setup so that I can tell if the creatuser form has been
submitted or if the form has not been submitted and we are seeing this
for the first time.

index.php - Include the code for the new createUser($post) function.